+++ SheetyColors/Classes/Common/Factory/SheetyColorsViewFactory.swift
@@ -10,14 +10,15 @@ import UIKit
 struct SheetyColorsViewFactory {
     static func createView(withConfig config: SheetyColorsConfigProtocol) -> SheetyColorsViewController {
         var viewModel: SheetyColorsViewModelProtocol
+        let hasTextOrMessage: Bool = config.title != nil || config.message != nil
 
         switch config.type {
         case .grayscale:
-            viewModel = GrayscaleViewModel(withColorModel: config.initialColor.grayscaleColor, alphaEnabled: config.alphaEnabled)
+            viewModel = GrayscaleViewModel(withColorModel: config.initialColor.grayscaleColor, isAlphaEnabled: config.alphaEnabled, hasTextOrMessage: hasTextOrMessage)
         case .hsb:
-            viewModel = HSBViewModel(withColorModel: config.initialColor.hsbaColor, alphaEnabled: config.alphaEnabled)
+            viewModel = HSBViewModel(withColorModel: config.initialColor.hsbaColor, isAlphaEnabled: config.alphaEnabled, hasTextOrMessage: hasTextOrMessage)
         case .rgb:
-            viewModel = RGBViewModel(withColorModel: config.initialColor.rgbaColor, alphaEnabled: config.alphaEnabled)
+            viewModel = RGBViewModel(withColorModel: config.initialColor.rgbaColor, isAlphaEnabled: config.alphaEnabled, hasTextOrMessage: hasTextOrMessage)
         }
 
         let view = SheetyColorsViewController.create()